Removes mixin
