--- lib/msf/core/post/dns.rb
@@ -1,4 +1,0 @@
-# -*- coding: binary -*-
-
-module Msf::Post::DNS
-end
